move query functions to client module
